feat: refactor index management for management users, permissions, sessions, and service user API keys
